+++ pkg/processor/func_wrapper_test.go
